Create file structure and separate model ts file for interfaces
