+++ src/model.ts
@@ -1,0 +1,5 @@
+export interface Joke {
+  id: string;
+  joke: string;
+  status: number;
+}
